get dose on load
